+++ tests/scene.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 use test_log::test;
 
 #[test]
-fn replicated_entity() {
+fn replicated() {
     let mut app = App::new();
     app.add_plugins((StatesPlugin, RepliconPlugins))
         .register_type::<TestComponent>()
@@ -14,7 +14,7 @@ fn replicated_entity() {
         .replicate::<NonReflectedComponent>()
         .finish();
 
-    let entity = app
+    let replicated = app
         .world_mut()
         .spawn((
             Replicated,
@@ -23,24 +23,36 @@ fn replicated_entity() {
             NonReflectedComponent,
         ))
         .id();
+    let remote = app.world_mut().spawn((Remote, TestComponent)).id();
 
     let mut scene = DynamicScene::default();
     scene::replicate_into(&mut scene, app.world());
 
-    assert!(scene.resources.is_empty());
-    assert_eq!(scene.entities.len(), 1);
+    let replicated_dyn = scene
+        .entities
+        .iter()
+        .find(|entity| entity.entity == replicated)
+        .unwrap();
 
-    let dyn_entity = &scene.entities[0];
-    assert_eq!(dyn_entity.entity, entity);
+    let remote_dyn = scene
+        .entities
+        .iter()
+        .find(|entity| entity.entity == remote)
+        .unwrap();
+
+    assert_eq!(replicated_dyn.entity, replicated);
     assert_eq!(
-        dyn_entity.components.len(),
+        replicated_dyn.components.len(),
         1,
         "entity should have only registered components with `#[reflect(Component)]`"
     );
+
+    assert_eq!(remote_dyn.entity, remote);
+    assert_eq!(remote_dyn.components.len(), 1);
 }
 
 #[test]
-fn empty_entity() {
+fn empty() {
     let mut app = App::new();
     app.add_plugins((StatesPlugin, RepliconPlugins)).finish();
 
@@ -59,7 +71,7 @@ fn empty_entity() {
 }
 
 #[test]
-fn not_replicated_entity() {
+fn not_replicated() {
     let mut app = App::new();
     app.add_plugins((StatesPlugin, RepliconPlugins))
         .register_type::<TestComponent>()
@@ -76,7 +88,7 @@ fn not_replicated_entity() {
 }
 
 #[test]
-fn entity_update() {
+fn update_existing() {
     let mut app = App::new();
     app.add_plugins((StatesPlugin, RepliconPlugins))
         .register_type::<TestComponent>()